% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noBio_accounting_functions.R
\name{allocate_biomass_emiss_reductions}
\alias{allocate_biomass_emiss_reductions}
\title{allocate_biomass_emiss_reductions}
\usage{
allocate_biomass_emiss_reductions(
  energy.consumption.by.tech,
  service.output.by.tech,
  market.prices,
  emiss.by.tech,
  bio.sectors
)
}
\arguments{
\item{energy.consumption.by.tech}{rgcam output from energy consumption by tech query. Columns must be 'Units', 'scenario', 'region', 'sector', 'subsector', 'technology', 'input', 'year', and 'value'}

\item{service.output.by.tech}{rgcam output from output by tech query. Columns must be 'Units', 'scenario', 'region', 'sector', 'subsector', 'technology', 'output', 'year', and 'value'}

\item{market.prices}{rgcam output from prices of all markets query. Columns must be 'Units', 'scenario', 'year', 'market', and 'value'}

\item{emiss.by.tech}{rgcam output from CO2 emissions by tech query. Columns must be 'Units', 'scenario', 'region', 'sector', 'subsector', 'technology', 'ghg', 'year', and 'value'}

\item{bio.sectors}{biomass sectors to account for (e.g. 'regional biomass')}
}
\value{
dataframe of CO2 emissions by sector with biomass negative emissions accounted for
}
\description{
perform full biomass accounting (transform co2 emissions by sector into co2 emissions by sector (no bio))
}
\author{
TRW
}
