% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noBio_accounting_functions.R
\name{apply_coefs}
\alias{apply_coefs}
\title{apply_coefs}
\usage{
apply_coefs(curr.FQ.sector, curr.emiss, coefs)
}
\arguments{
\item{curr.FQ.sector}{the sector to start with (a biomass sector including the region name, such as 'Thailandregional biomass')}

\item{curr.emiss}{CO2 emissions by technology data filtered to a given scenario, region, and biomass sector}

\item{coefs}{list of carbon stay/go coefficients for each input}
}
\value{
list of amounts of emissions staying in and leaving from each sector
}
\description{
apply carbon in/out coefficients recursively through sectors for a given input
}
\author{
TRW
}
