% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assumptions.R
\name{assumptions}
\alias{assumptions}
\title{assumptions}
\usage{
assumptions(name = NULL)
}
\arguments{
\item{name}{Default=NULL. Name of assumption object.
List of Assumptions
\itemize{
\item GCAMbaseYear,
\item convEJ2MTOE,
\item convEJ2TWh,
\item convEJ2GW,
\item convEJ2GWh,
\item convGW_kW,
\item conv_C_CO2,
\item conv_MT_GT,
\item hydro_cap_fact,
\item hydro_cost_GW,
\item convUSD_1975_2010,
\item convUSD_1975_2015,
\item conv1975USDperGJ22017USDperMWh,
\item conv1975USDperGJ22017USDperMBTU,
\item convertGgTgMTC,
\item GWP,
\item US52,
\item US49,
\item countryToGCAMReg32}}
}
\value{
A list of assumptions
}
\description{
This function loads holds the different assumptions used throughout the gcamextractor package.
}
\examples{
library(gcamextractor)
a<-assumptions("US49")
}
\keyword{assumptions}
