% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capfactors}
\alias{capfactors}
\title{capfactors}
\format{
.csv
}
\source{
paste(getwd(),"/dataFiles/gcam/capacity_factor_by_elec_gen_subsector.csv",sep="")
}
\usage{
capfactors
}
\description{
capfactors
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::capfactors
}
}
\keyword{datasets}
