% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conv_GgTg_to_MTC}
\alias{conv_GgTg_to_MTC}
\title{conv_GgTg_to_MTC}
\format{
tibble
}
\source{
gcamextractor
}
\usage{
conv_GgTg_to_MTC
}
\description{
https://nepis.epa.gov/Exe/ZyNET.exe/P1001YTS.txt?ZyActionD=ZyDocument&Client=EPA&Index=2000%20Thru%202005&Docs=&Query=&Time=&EndTime=&SearchMethod=1&TocRestrict=n&Toc=&TocEntry=&QField=&QFieldYear=&QFieldMonth=&QFieldDay=&UseQField=&IntQFieldOp=0&ExtQFieldOp=0&XmlQuery=&File=D%3A%5CZYFILES%5CINDEX%20DATA%5C00THRU05%5CTXT%5C00000017%5CP1001YTS.txt&User=ANONYMOUS&Password=anonymous&SortMethod=h%7C-&MaximumDocuments=1&FuzzyDegree=0&ImageQuality=r75g8/r75g8/x150y150g16/i425&Display=hpfr&DefSeekPage=x&SearchBack=ZyActionL&Back=ZyActionS&BackDesc=Results%20page&MaximumPages=1&ZyEntry=3
https://www.firescience.gov/projects/09-2-01-9/supdocs/09-2-01-9_Appendix_C_-_Unit_Conversion_and_Other_Tables.pdf
MTC is megatonnes (10^6 tonnes) of Carbon.
Tg one terragram = 1 Megatonne = 10^6 tonnes
Convert to Mega tonnes of Carbon
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::conv_GgTg_to_MTC
}
}
\keyword{datasets}
