% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{convert}
\alias{convert}
\title{convert}
\format{
list
}
\source{
gcamextractor
}
\usage{
convert
}
\description{
convert
}
\examples{
\dontrun{
 library(gcamextractor);
 names(gcamextractor::convert)
}
}
\keyword{datasets}
