% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_GWP}
\alias{data_GWP}
\title{data_GWP}
\format{
tibble
}
\source{
gcamextractor
}
\usage{
data_GWP
}
\description{
Emissions Conversion to CO2eq
GWP conversions - uses 100-yr GWPs from IPPC AR4 and AR5
https://www.ghgprotocol.org/sites/default/files/ghgp/Global-Warming-Potential-Values%20%28Feb%2016%202016%29_1.pdf
Does not include all covnersions. Add them if they are tracked in GCAM
Mean HFC numbers GWP %>% dplyr::filter(grepl("HFC",ghg)) %>% summarize_at(vars(names(GWP)[!grepl("ghg",names(GWP))]),funs(mean))
GTP AR5 Box 3.2 Table 1 https://www.ipcc.ch/site/assets/uploads/2018/02/SYR_AR5_FINAL_full.pdf
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::data_GWP
}
}
\keyword{datasets}
