% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_params}
\alias{data_params}
\title{gcamextractor params}
\format{
R table
}
\source{
from gcamextractor::mappings()$mapParamQuery$param
}
\usage{
data_params
}
\description{
gcamextractor params
}
\examples{
\dontrun{
 library(gcmaextractor);
 gcmaextractor::data_params
}
}
\keyword{datasets}
