% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_queries}
\alias{data_queries}
\title{gcamextractor queries}
\format{
R table
}
\source{
from gcamextractor::mappings()$mapParamQuery$query
}
\usage{
data_queries
}
\description{
gcamextractor queries
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::data_queries
}
}
\keyword{datasets}
