% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_GCAMv52_2050_proj}
\alias{example_GCAMv52_2050_proj}
\title{Example GCAMv5.2 .proj file}
\format{
proj file
}
\source{
readgcam() run outputs saved
}
\usage{
example_GCAMv52_2050_proj
}
\description{
Example GCAMv5.2 .proj file
}
\examples{
\dontrun{
 library(gcmaextractor);
 gcmaextractor::example_GCAMv52_2050_proj
}
}
\keyword{datasets}
