% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_gcamv54_argentina_colombia_2025_proj}
\alias{example_gcamv54_argentina_colombia_2025_proj}
\title{example_gcamv54_argentina_colombia_2025_proj}
\format{
proj file
}
\source{
readgcam() run outputs saved
}
\usage{
example_gcamv54_argentina_colombia_2025_proj
}
\description{
example_gcamv54_argentina_colombia_2025_proj
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::example_gcamv54_argentina_colombia_2025
}
}
\keyword{datasets}
