% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleGCAMproj}
\alias{exampleGCAMproj}
\title{Example GCAM .proj file}
\format{
R table or .csv
}
\source{
readgcam() run outputs saved
}
\usage{
exampleGCAMproj
}
\description{
Example GCAM .proj file
}
\examples{
\dontrun{
 library(gcmaextractor);
 gcmaextractor::exampleGCAMproj
}
}
\keyword{datasets}
