% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcamextractor.R
\docType{package}
\name{gcamextractor}
\alias{gcamextractor}
\title{gcamextractor}
\description{
\itemize{
\item Github: https://github.com/JGCRI/gcamextractor
\item Webpage: https://jgcri.github.io/gcamextractor/}
}
