% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gcamextractor_test_data}
\alias{gcamextractor_test_data}
\title{gcamextractor_test_data
Data for testing}
\format{
tibble
}
\source{
gcamextractor_test_data
}
\usage{
gcamextractor_test_data
}
\description{
gcamextractor_test_data
Data for testing
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::gcamextractor_test_data
}
}
\keyword{datasets}
