% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noBio_accounting_functions.R
\name{generate_sector_output_coefs}
\alias{generate_sector_output_coefs}
\title{generate_sector_output_coefs}
\usage{
generate_sector_output_coefs(curr.FQinput, data)
}
\arguments{
\item{curr.FQinput}{the energy input to calculate carbon in/out proportions for}

\item{data}{output from prepare_sector_carbon_flows filtered to one scenario}
}
\value{
list of carbon in/out proportions for each sector
}
\description{
Make a list of carbon in/out proportions across sectors for a given input
}
\author{
TRW
}
