% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_data.R
\name{get_example_data}
\alias{get_example_data}
\title{get_example_data}
\usage{
get_example_data(
  write_dir = getwd(),
  dir_name = "gcamextractor_example_data",
  data_link = NULL
)
}
\arguments{
\item{write_dir}{Default = getwd()}

\item{dir_name}{= "gcamextractor_example_data". Name of directory to install zip file to.}

\item{data_link}{Default = NULL}
}
\value{
number
}
\description{
Function to download the test data files from zenodo that are needed.
}
\examples{
\dontrun{
library(osiris)
osiris::get_example_data()
}
}
\keyword{test}
