% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_country_to_gcam_region}
\alias{map_country_to_gcam_region}
\title{map_country_to_gcam_region}
\format{
tibble
}
\source{
gcamextractor
}
\usage{
map_country_to_gcam_region
}
\description{
map_country_to_gcam_region
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::map_country_to_gcam_region
}
}
\keyword{datasets}
