% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_param_query}
\alias{map_param_query}
\title{map_param_query}
\format{
tibble
}
\source{
gcamextractor
}
\usage{
map_param_query
}
\description{
map_param_query
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::map_param_query
}
}
\keyword{datasets}
