% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map_state_to_gridregion}
\alias{map_state_to_gridregion}
\title{map_state_to_gridregion}
\format{
tibble
}
\source{
gcamextractor
}
\usage{
map_state_to_gridregion
}
\description{
map_state_to_gridregion
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::map_state_to_gridregion
}
}
\keyword{datasets}
