% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params}
\alias{params}
\title{gcamextractor params}
\format{
R table
}
\source{
from gcamextractor::mappings()$mapParamQuery$param
}
\usage{
params
}
\description{
gcamextractor params
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::params
}
}
\keyword{datasets}
