% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noBio_accounting_functions.R
\name{prepare_sector_carbon_flows}
\alias{prepare_sector_carbon_flows}
\title{prepare_sector_carbon_flows}
\usage{
prepare_sector_carbon_flows(
  energy.consumption.by.tech,
  service.output.by.tech,
  market.prices
)
}
\arguments{
\item{energy.consumption.by.tech}{rgcam output from energy consumption by tech query. Columns must be 'Units', 'scenario', 'region', 'sector', 'subsector', 'technology', 'input', 'year', and 'value'}

\item{service.output.by.tech}{rgcam output from output by tech query. Columns must be 'Units', 'scenario', 'region', 'sector', 'subsector', 'technology', 'output', 'year', and 'value'}

\item{market.prices}{rgcam output from prices of all markets query. Columns must be 'Units', 'scenario', 'year', 'market', 'value'}
}
\value{
dataframe of carbon inputs and outputs
}
\description{
Calculate proportions of carbon in and out for each sector and energy input
}
\author{
TRW
}
