% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queries_xml}
\alias{queries_xml}
\title{queries xml file}
\format{
.xml
}
\source{
gcamextractor
}
\usage{
queries_xml
}
\description{
queries xml file
}
\examples{
\dontrun{
 library(gcamextractor); library(XML)
 gcamextractor::queries_xml
 # Can save xml
 XML::saveXML(gcamextractor::queries_xml, file=paste(getwd(), "/queries.xml", sep = ""))
}
}
\keyword{datasets}
