% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions_gcam32}
\alias{regions_gcam32}
\title{regions_gcam32}
\format{
tibble
}
\source{
gcamextractor
}
\usage{
regions_gcam32
}
\description{
regions_gcam32
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::regions_gcam32
}
}
\keyword{datasets}
