% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions_US49}
\alias{regions_US49}
\title{regions_US49}
\format{
tibble
}
\source{
gcamextractor
}
\usage{
regions_US49
}
\description{
# GCAM USA 49. Excludes Alaska, Hawaii and Puerto Rico. Includes DC.
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::regions_US49
}
}
\keyword{datasets}
