% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regions_US52}
\alias{regions_US52}
\title{regions_US52}
\format{
tibble
}
\source{
gcamextractor
}
\usage{
regions_US52
}
\description{
regions_US52
}
\examples{
\dontrun{
 library(gcamextractor);
 gcamextractor::regions_US52
}
}
\keyword{datasets}
