% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{queries}
\alias{queries}
\title{queries xml file}
\format{
.xml
}
\source{
gcamextractor
}
\usage{
queries
}
\description{
queries xml file
}
\examples{
\dontrun{
 library(gcamextractor); library(XML)
 gcamextractor::queries
 # Can save xml
 XML::saveXML(gcamextractor::queries, file=paste(getwd(), "/queries.xml", sep = ""))
}
}
\keyword{datasets}
